+++ RELEASE_NOTES.md
@@ -20,6 +20,7 @@ The format of this file is based on [Keep a Changelog](http://keepachangelog.com
 - Adds missing synchronize when using pinned memory.
 
 ### Removed
+- Removes ManagedArray::getActiveBasePointer method.
 - Removes deprecated ManagedArray::getPointer method. Use ManagedArray::data instead.
 - Removes ManagedArray::incr and ManagedArray::decr methods. Use ManagedArray::pick and ManagedArray::set instead.
 - Removes optional support for implicitly casting between raw pointers and ManagedArrays (CHAI\_ENABLE\_IMPLICIT\_CONVERSIONS). Use makeManagedArray and ManagedArray::data to perform explicit conversions instead.